separated user creation logic and mqtt credential creation logic call
