+++ src/main/java/com/logreposit/logrepositapi/services/mqtt/MqttCredentialService.java
@@ -17,6 +17,4 @@ public interface MqttCredentialService {
       throws MqttCredentialServiceException;
 
   MqttCredential getGlobalDeviceDataWriteCredential();
-
-  boolean isMqttSupportEnabled();
 }
